feat: Added separate trends section and new graph for tickets created vs closed
